REF: move code examples to composables
